Add files via upload
The recent pandemic situation has worsened the unemployment scenario, and people are struggling to find employment opportunities. That is where my Android application comes in - it provides a platform that connects job providers with job seekers, making the entire process of hiring much simpler.
The application is built using Firebase and Java, which are two of the most popular and powerful technologies available in the market. Firebase provides a secure and reliable backend infrastructure for the application, while Java allows for the creation of beautiful and efficient user interfaces. Together, they make for a great combination that allows for the development of robust and feature-rich applications.
FitMyJob is an android application which acts as a portal between Job Seekers and Recruiters. Using FitMyJob one can search job matching his or her requirements and appear for the selection process once the application is being registered and accepted. It can be very helpful as it allows rural side different skilled people to upload their details and search matching job according to their skills and interest. Also, every user can apply for multiple job at a time.
FitMyJob brings candidate and top employees under one roof. While most online job portals focus on getting candidates the next job, FitMyJob focuses on the entire career growth of the rural side skilled people